fix: configure DHIS2 component internal queue to be unbounded in order to eliminate errors caused by large number of splits
rename packages
