+++ src/main/java/org/hisp/dhis/integration/t2a/processor/AnalyticsGridQueryBuilder.java
@@ -1,0 +1,57 @@
+/*
+ * Copyright (c) 2004-2022, University of Oslo
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ * Redistributions of source code must retain the above copyright notice, this
+ * list of conditions and the following disclaimer.
+ *
+ * Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ * Neither the name of the HISP project nor the names of its contributors may
+ * be used to endorse or promote products derived from this software without
+ * specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
+ * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package org.hisp.dhis.integration.t2a.processor;
+
+import java.util.List;
+import java.util.Map;
+
+import org.apache.camel.Exchange;
+import org.apache.camel.Processor;
+import org.hisp.dhis.integration.t2a.model.Dimensions;
+import org.hisp.dhis.integration.t2a.route.T2ARouteBuilder;
+import org.springframework.stereotype.Component;
+
+@Component
+public class AnalyticsGridQueryBuilder implements Processor
+{
+    public void process( Exchange exchange )
+    {
+        Dimensions dimensions = exchange.getMessage().getBody( Dimensions.class );
+
+        Map<String, Object> query = Map.of( "dimension",
+            List.of( "dx:" + dimensions.getProgramIndicator().getId().get(),
+                "ou:" + dimensions.getOrganisationUnitIds(), "pe:" + dimensions.getPeriods() ),
+            "rows",
+            "ou;pe",
+            "columns", "dx", "skipMeta", "true" );
+
+        exchange.setProperty( T2ARouteBuilder.DIMENSIONS_PROPERTY, dimensions );
+        exchange.getMessage()
+            .setHeader( "CamelDhis2.queryParams", query );
+    }
+}
