+++ src/main/java/org/hisp/dhis/integration/t2a/route/T2ARouteBuilder.java
@@ -1,0 +1,72 @@
+/*
+ * Copyright (c) 2004-2022, University of Oslo
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ * Redistributions of source code must retain the above copyright notice, this
+ * list of conditions and the following disclaimer.
+ *
+ * Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ * Neither the name of the HISP project nor the names of its contributors may
+ * be used to endorse or promote products derived from this software without
+ * specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
+ * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package org.hisp.dhis.integration.t2a.route;
+
+import org.apache.camel.LoggingLevel;
+import org.apache.camel.builder.RouteBuilder;
+import org.apache.camel.builder.ThreadPoolProfileBuilder;
+import org.apache.camel.spi.ThreadPoolProfile;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Component;
+
+/**
+ * Pulling program indicators one at a time as DataValueSets (beware if your
+ * pool size is too big it breaks the database :-( )
+ */
+@Component
+public class T2ARouteBuilder extends RouteBuilder
+{
+    private static final Logger LOGGER = LoggerFactory.getLogger( T2ARouteBuilder.class );
+
+    public static final String DIMENSIONS_PROPERTY = "dimensions";
+
+    public static final String ALL_ORG_UNITS_PROPERTY = "ous";
+
+    @Override
+    public void configure()
+    {
+        ThreadPoolProfile camelDhis2ThreadProfile = new ThreadPoolProfileBuilder( "CamelDhis2" ).build();
+        camelDhis2ThreadProfile.setMaxQueueSize( -1 );
+        getContext().getExecutorServiceManager().registerThreadPoolProfile( camelDhis2ThreadProfile );
+
+        from( "seda:t2a" ).setProperty( "startTime", simple( "${bean:java.lang.System?method=currentTimeMillis}" ) )
+            .streamCaching( "true" )
+            .log( LoggingLevel.INFO, LOGGER, "Fetching organisation units..." )
+            .toD(
+                "dhis2://get/resource?path=organisationUnits&fields=id&filter=level:eq:{{org.unit.level}}&client=#dhis2Client" )
+            .setProperty( ALL_ORG_UNITS_PROPERTY, jsonpath( "$.organisationUnits..id" ) )
+            .setHeader( "skipAggregate", constant( "true" ) )
+            .setHeader( "skipEvents", constant( "false" ) )
+            .choice().when( simple( "{{run.event.analytics:true}}" ) ).to( "direct:run-analytics" ).end()
+            .to( "direct:collect" )
+            .process( e -> e.getIn()
+                .setBody( (System.currentTimeMillis() - e.getProperty( "startTime", Long.class )) / 1000 ) )
+            .log( LoggingLevel.INFO, LOGGER, "Aggregated program indicators in ${body} seconds" );
+    }
+}
