+++ src/test/java/org/hisp/dhis/integration/t2a/route/PushRouteBuilderTestCase.java
@@ -1,0 +1,85 @@
+/*
+ * Copyright (c) 2004-2022, University of Oslo
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ * Redistributions of source code must retain the above copyright notice, this
+ * list of conditions and the following disclaimer.
+ *
+ * Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ * Neither the name of the HISP project nor the names of its contributors may
+ * be used to endorse or promote products derived from this software without
+ * specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
+ * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package org.hisp.dhis.integration.t2a.route;
+
+import static org.hisp.dhis.integration.t2a.route.T2ARouteBuilder.ALL_ORG_UNITS_PROPERTY;
+import static org.junit.jupiter.api.Assertions.assertNull;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
+import org.apache.camel.Exchange;
+import org.apache.camel.ProducerTemplate;
+import org.apache.camel.support.DefaultExchange;
+import org.apache.camel.test.spring.junit5.CamelSpringBootTest;
+import org.apache.camel.test.spring.junit5.UseAdviceWith;
+import org.hisp.dhis.api.v2_37_6.model.ProgramIndicator;
+import org.hisp.dhis.api.v2_37_6.model.ProgramIndicatorGroup;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+@CamelSpringBootTest
+@UseAdviceWith
+public class PushRouteBuilderTestCase extends AbstractRouteBuilderTestCase
+{
+    @Autowired
+    protected ProducerTemplate producerTemplate;
+
+    @BeforeEach
+    public void beforeEach()
+        throws Exception
+    {
+        System.setProperty( "org.unit.batch.size",
+            String.valueOf( 1 ) );
+        super.beforeEach();
+
+    }
+
+    @Test
+    public void testPushRouteDoesNotExceedQueueSizeLimitWhenGivenLargeNoOfOrgUnits()
+    {
+        List<String> orgUnitIds = new ArrayList<>();
+        for ( int i = 0; i < 5000; i++ )
+        {
+            orgUnitIds.add( orgUnitUnderTestId );
+        }
+
+        Exchange exchange = new DefaultExchange( camelContext );
+        exchange.setProperty( ALL_ORG_UNITS_PROPERTY, orgUnitIds );
+        exchange.getMessage().setBody(
+            new ProgramIndicatorGroup().withProgramIndicators( List.of( new ProgramIndicator().withAttributeValues(
+                Collections.emptyList() ).withId( "yC212U3ifgY" ) ) ) );
+
+        assertNull( producerTemplate.send( "direct:push", exchange ).getException() );
+    }
+
+}
